Implement core of option_b approach
Implement the ControlledCacheWarmer, which prevents the
ContainerChecker from returning a pod unless it is warm (i.e.,
in the cache).  If it is not, the warmer runs a thread that
attempts to warm all active pods, one at a time.

TODO: Hook in Ed's token-based approach for accessing the API.
TODO: Reshape K8s temporary errors to all throw a K8sApiTemporaryError exception.
TODO: Implement retries on the path to fetch a pod and its controller info.